Add files via upload
Question 1 – API Testing - Instructions:
You will need to build a Postman Collection to test the API. 
Question 1:
• If you go to https://pos.com.my/send/ratecalculator -- you will find a Shipping
Rate Calculator functionality.
Find the API that lists all the countries available in the “To” dropdown and create
the request (1 request).
• Enter the postcode in the From and To. The states will be automatically
populated. Find the API that’s responding to the postcode number being entered
by the user and create the request (1 request).
• Enter the complete From, To, and Weight, and click Calculate. You will get a
quote. Retrieve the API that calculates the rate and build the equivalent
Postman request.(1 request)